Patch the grammar to disambiguate matching of keywords in simple-labels
Implements https://github.com/dhall-lang/dhall-lang/pull/299

We add a negative lookahead in the simple-label rule, so that
if a label starts with a keyword we backtrack.
(Note: there's a space added, to avoid refusing partial matches,
e.g. if a keyword is "in" we should still accept "index" as a label.